Changed name of page to outfits
Changed the name to outfits from about .
