+++ ionic.config.json
@@ -1,6 +1,6 @@
 {
-  "name": "Ionic2FirebaseEmail",
-  "app_id": "",
+  "name": "Textile",
+  "app_id": "3a4b03cd",
   "v2": true,
   "typescript": true
 }
