--- src/pages/about/about.html
@@ -1,45 +1,0 @@
-
-<ion-header>
-  <ion-navbar hideBackButton="true">
-    <ion-buttons >
-		<button color="navcolor" ion-button navPop icon-only>
-			<ion-icon ios="ios-arrow-back" md="md-arrow-back"></ion-icon>
-		</button>
-	</ion-buttons>
-   <ion-buttons end>
-      <button ion-button icon-only color="primary"  (click)="enableDelete()" >
-        <ion-icon color="navcolor" ios="ios-trash-outline" md="md-trash"></ion-icon>
-      </button>
- </ion-buttons>
-    <ion-title>
-     <p style="font-size:22px;">Outfits</p>
-    </ion-title>
-  </ion-navbar>
-</ion-header>
-<ion-content>
-
-   <ion-row wrap>
-       
-
-       
-  <ion-col width-33 *ngFor="let f of outfits">
-    
-
-    <ion-card no-padding> 
-        <div style="position: relative; top:0%; left:0px;">
-    <ion-fab right top>
-      <button *ngIf ="isEnabled"ion-fab mini class="fab-map" mini color="navcolor" (click)="deleteOutfits(f)">
-        <ion-icon name="close"></ion-icon>
-      </button>
-    </ion-fab>
-  </div>
-  <img *ngIf="f.first"[src]="f.first">
-  <img *ngIf="f.second"[src]="f.second">
-    
-
-  <img *ngIf="f.third"[src]="f.third">
-  <img *ngIf="f.fourth"[src]="f.fourth">
-    </ion-card>
-  </ion-col>
-</ion-row>
-</ion-content>
