--- src/pages/about/about.scss
@@ -1,15 +1,0 @@
-page-about {
-
-}
-.fab[mini] {
-    margin: 0px;
-    width: 11px;
-    height: 28px;
-    line-height: 0px;
-}
-
-button[fab].fab-map {
-  position: absolute;
- 
-  z-index: 100;
-}
